+++ src/BigCommerce/PriceLists/PriceListRecordsApi.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace BigCommerce\ApiV3\PriceLists;
+
+class PriceListRecordsApi
+{
+
+}