+++ src/BigCommerce/PriceLists/PriceListsApi.php
@@ -3,25 +3,23 @@
 namespace BigCommerce\ApiV3\PriceLists;
 
 use BigCommerce\ApiV3\Api\ResourceApi;
-use BigCommerce\ApiV3\ResponseModels\PaginatedResponse;
 use BigCommerce\ApiV3\ResponseModels\PriceList\PriceListResponse;
 use BigCommerce\ApiV3\ResponseModels\PriceList\PriceListsResponse;
-use BigCommerce\ApiV3\ResponseModels\SingleResourceResponse;
 
 class PriceListsApi extends ResourceApi
 {
     private const RESOURCE_NAME   = 'pricelists';
-    private const BRAND_ENDPOINT  = 'pricelists/%d';
-    private const BRANDS_ENDPOINT = 'pricelists';
+    private const PRICELIST_ENDPOINT  = 'pricelists/%d';
+    private const PRICELISTS_ENDPOINT = 'pricelists';
 
     protected function singleResourceEndpoint(): string
     {
-        return self::BRAND_ENDPOINT;
+        return self::PRICELIST_ENDPOINT;
     }
 
     protected function multipleResourcesEndpoint(): string
     {
-        return self::BRANDS_ENDPOINT;
+        return self::PRICELISTS_ENDPOINT;
     }
 
     protected function resourceName(): string
@@ -37,5 +35,10 @@ class PriceListsApi extends ResourceApi
     public function getAll(array $filters = [], int $page = 1, int $limit = 250): PriceListsResponse
     {
         return new PriceListsResponse($this->getAllResources($filters, $page, $limit));
+    }
+
+    public function assignments(): PriceListAssignmentsApi
+    {
+        return new PriceListAssignmentsApi($this->getClient(), null, $this->getResourceId());
     }
 }