+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 
 Adds support for the [Price Lists API](https://developer.bigcommerce.com/api-reference/catalog/pricelists-api)
 
+Includes:
+ - Price List Assignments